config: Endless customizations
Build in autofs4 support, to avoid the initramfs complaining that
autofs4 is not available.

PINCTRL_SUNRISEPOINT needs to be built in to make the volume control
buttons on the side of the Asux UX360UA work correctly. (T13917)

PINCTRL_BROXTON needs to be built in otherwise sdhci-pci cannot detect
the SD card on several Asus ApolloLake models e.g. X441NA. (T14845)

Enable RTL8192CU vendor driver and disable the mainline one.

Enable several other drivers that we add.

Enable DRM_BOCHS, the default VGA card emulated by qemu.
The Ubuntu issue which led to this being disabled --
https://bugs.launchpad.net/ubuntu/+source/kmod/+bug/1378648 --
seems to be ppc64el-specific, so not relevant to us. (T19074)