RTL8723BE: Fix antenna selection for HP laptops
The number of antennas connected to the WiFi/BT transceiver and their
actual connector is an information carried inside the on-board EEPROM.
When this information is wrong because the EEPROM is incorrectly
programmed, the WiFi/BT signal is really weak and a stable connection is
practically impossible.

The EEPROM on the HP laptop correctly indicates that 1 single antenna
is present (ant_num == ANT_X1 == 1) but the driver by default assumes
that this is connected to BTC_ANTENNA_AT_MAIN_PORT. In our case (found
experimentally) the single antenna is connected instead to
BTC_ANTENNA_AT_AUX_PORT.

Acting on the 'ant_num' module parameter we can discard the information
carried by the EEPROM and force the driver to consider the single
antenna connected to the BTC_ANTENNA_AT_AUX_PORT (using 'ant_num == 1').

Readings:
https://patchwork.kernel.org/patch/8603231/
https://patchwork.kernel.org/patch/8603241/

We have seen two instances of the 240 G5 model, one with
DMI vendor "HP" and another with "Hewlett-Packard". In order to apply
the workaround in both cases, we do not filter by the DMI vendor string.

https://phabricator.endlessm.com/T15935
https://phabricator.endlessm.com/T15935
https://phabricator.endlessm.com/T20141

Signed-off-by: Carlo Caione <carlo@endlessm.com>