rtl8xxxu: Disable binding for 0x0bda:0xb720 (rtl8723bu)
The Zyrex Sky laptop is shipping the 0x0bda:0xb720 WiFi transceiver. We
want to avoid using the upstream driver for this module since it is not
working fine. This is not the only hardware affected by this problem, so
instead of adding a quirk only for the Zyrex laptop we disable the
binding on the whole VID:PID.

The downstream driver is already in place in
drivers/net/wireless/realtek/rtl8723bu and we want to use that for this
hardware.

Signed-off-by: Carlo Caione <carlo@endlessm.com>